core/state: implement binary hasher just for demonstration
